adding data from the project
